+++ Editor/ValidationSuite/ValidationTests/ApiValidation.cs
@@ -17,7 +17,7 @@ namespace UnityEditor.PackageManager.ValidationSuite.ValidationTests
             TestName = "API Validation";
             TestDescription = "Checks public API for style and changest that conflict with Semantic Versioning.";
             TestCategory = TestCategory.ApiValidation;
-            SupportedValidations = new[] { ValidationType.CI, ValidationType.LocalDevelopmentInternal, ValidationType.Publishing };
+            SupportedValidations = new[] { ValidationType.CI, ValidationType.LocalDevelopmentInternal, ValidationType.Promotion };
             CanUseValidationExceptions = true;
         }